requirements.txt
opencv-python
onvif-zeep
requests
pillow
pyyaml
numpy<2.0 